src/requirements.txt
matplotlib
numpy 
pandas
scipy
pyyaml
